// src/app/api/generated/models/tour-log-dto.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { TourLogWeatherDto } from '../models/tour-log-weather-dto';
export interface TourLogDto {
  comment?: string;
  createdAt?: string;
  difficulty?: number;
  id?: number;
  performedAt?: string;
  rating?: number;
  totalDistanceM?: number;
  totalTimeS?: number;
  tourId?: number;
  updatedAt?: string;
  version?: number;
  weather?: TourLogWeatherDto;
}